main: Add helpers to compute other statistics
This involves predicting the next k words from the model

The way this works is a bit interesting, since the model itself
is quite heavily optimized for training and not ergonomics. Since
it is trying to predict an array of the i + 1th word for each of the
i words in each sentence for each of the b batches, we need to:

1) Set the batch size to 1
2) Set the hidden layer height to 1

That way, we only make a single prediction and we can just iteratively
predict using the last word and last hidden state.

We then convert back into words, stick the words into a sentence
and then apply them metrics.