model: Add support for shortlists
Pass --shortlist with some length to the command line to reduce
the size of the output layers to only predict the top N most frequent words.

The way that this works is that we remap the training targets to
the "shortlisted" words, remapping anything that is not common enough
to the unknown token and anything else getting a pass-through remapping.

At training time we train the model to predict UNK for the uncommon
words. At validation time we compare the actual sentences in the training
data (without the remapping) to the predictions made by the model (which
will have the remapping) and compute accuracy from there.